fix: clean up volumes through CI (#127)
Found that I had this clean up script already in the repo, so I'm adding
it to the repo officially.

I then call this script from the pipeline so it automatically cleans up
the volumes. However, I'm a bit worried it'll clean up all the volumes
and we may need one?