Enable Jaeger on AMI instances (#131)
Enables Jaeger by default for AMI instances. 

As the intent of the AMI is minimize configuration requirements below the web UI, deploying the Jaeger container in the AMI by default removes this step. It also results in the image getting baked into the AMI, so that customers who cannot pull images from Docker Hub can still use HTTP traces when needed to troubleshoot issues with their instance.

After / if this PR is merged, a corresponding update to the HTTP tracing docs will be made, to include the AMI
https://sourcegraph.com/docs/admin/observability/tracing

Tested manually on AWS AMI