chore(release): add tagging to the AMI build pipeline (#153)
This PR adds tagging the `deploy` repo at the time we build and release
new AMIs. I don't believe the code is changing much in this repo on
Sourcegraph version changes, but if it does this helps us keep track of
it.

This is tagging the branch opened when we create a PR to update the
Changelog, as the given version tag